move localizations to ABNotifier.bundle
By default, if the localizations are included directly in *.lproj directories
XCode will assume the application is localized for that locale. This also
causes the application to be displayed in the app store as supporting
localization for those locales. Moving the localizations to a bundle allows the
same drag/drop installation without having to worry about the consuming
application managing the localizations.